# openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Jikū API",
    "description" : "Event invitation, ticketing, RSVP and check-in API.",
    "version" : "v1"
  },
  "servers" : [
    {
      "url" : "http://localhost",
      "description" : "Generated server url"
    }
  ],
  "security" : [
    {
      "bearer-jwt" : [ ]
    }
  ],
  "paths" : {
...
    "/api/v1/events/{id}" : {
      "get" : {
        "tags" : [
          "event-controller"
        ],
        "operationId" : "get_1",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [
          "event-controller"
        ],
        "operationId" : "update_1",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateEventRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{id}/quorum" : {
      "put" : {
        "tags" : [
          "event-controller"
        ],
        "operationId" : "setQuorum",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateQuorumRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QuorumResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{eventId}/ticket-types/{typeId}" : {
      "put" : {
        "tags" : [
          "ticket-type-controller"
        ],
        "operationId" : "update_2",
        "parameters" : [
          {
            "name" : "eventId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          },
          {
            "name" : "typeId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpsertTicketTypeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketTypeResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [
          "ticket-type-controller"
        ],
        "operationId" : "delete",
        "parameters" : [
          {
            "name" : "eventId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          },
          {
            "name" : "typeId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
    "/api/v1/notifications/email-feedback" : {
      "post" : {
        "tags" : [
          "email-feedback-controller"
        ],
        "operationId" : "receive",
        "parameters" : [
          {
            "name" : "X-Webhook-Secret",
            "in" : "header",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EmailFeedbackWebhookRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "description" : "Accepted"
          }
        }
      }
    },
    "/api/v1/notifications/email-feedback/resend" : {
      "post" : {
        "tags" : [
          "resend-email-feedback-controller"
        ],
        "operationId" : "receive_1",
        "parameters" : [
          {
            "name" : "svix-id",
            "in" : "header",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "svix-timestamp",
            "in" : "header",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "svix-signature",
            "in" : "header",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "description" : "Accepted"
          }
        }
      }
    },
...
    "/api/v1/events" : {
      "get" : {
        "tags" : [
          "event-controller"
        ],
        "operationId" : "list",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/EventResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [
          "event-controller"
        ],
        "operationId" : "create_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateEventRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{id}/publish" : {
      "post" : {
        "tags" : [
          "event-controller"
        ],
        "operationId" : "publish",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{id}/cancel" : {
      "post" : {
        "tags" : [
          "event-controller"
        ],
        "operationId" : "cancel",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events/{eventId}/ticket-types" : {
      "get" : {
        "tags" : [
          "ticket-type-controller"
        ],
        "operationId" : "list_2",
        "parameters" : [
          {
            "name" : "eventId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TicketTypeResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [
          "ticket-type-controller"
        ],
        "operationId" : "create_3",
        "parameters" : [
          {
            "name" : "eventId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpsertTicketTypeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketTypeResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events/{eventId}/payments/manual" : {
      "get" : {
        "tags" : [
          "manual-payment-controller"
        ],
        "operationId" : "current",
        "parameters" : [
          {
            "name" : "eventId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ManualPaymentInstructions"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [
          "manual-payment-controller"
        ],
        "operationId" : "request",
        "parameters" : [
          {
            "name" : "eventId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ManualPaymentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ManualPaymentInstructions"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "EventSettingsDto" : {
        "type" : "object",
        "properties" : {
          "placementEnabled" : {
            "type" : "boolean"
          },
          "transferAllowed" : {
            "type" : "boolean"
          },
          "transferDeadline" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "date-time"
          },
          "overbookingAllowed" : {
            "type" : "boolean"
          },
          "maxOverbookingCount" : {
            "type" : [
              "integer",
              "null"
            ],
            "format" : "int32"
          }
        }
      },
      "UpdateEventRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "minLength" : 1
          },
          "description" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "startDateTime" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "date-time"
          },
          "endDateTime" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "date-time"
          },
          "timezone" : {
            "type" : "string",
            "minLength" : 1
          },
          "location" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "maxCapacity" : {
            "type" : [
              "integer",
              "null"
            ],
            "format" : "int32"
          },
          "settings" : {
            "$ref" : "#/components/schemas/EventSettingsDto"
          },
          "invitationChannels" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [
                "EMAIL",
                "WHATSAPP"
              ]
            },
            "uniqueItems" : true
          }
        },
        "required" : [
          "name",
          "timezone"
        ]
      },
      "EventResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "startDateTime" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "date-time"
          },
          "endDateTime" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "date-time"
          },
          "timezone" : {
            "type" : "string"
          },
          "location" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "maxCapacity" : {
            "type" : [
              "integer",
              "null"
            ],
            "format" : "int32"
          },
          "status" : {
            "type" : "string"
          },
          "settings" : {
            "$ref" : "#/components/schemas/EventSettingsDto"
          },
          "invitationChannels" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [
                "EMAIL",
                "WHATSAPP"
              ]
            },
            "uniqueItems" : true
          },
          "quorum" : {
            "oneOf" : [
              {
                "$ref" : "#/components/schemas/QuorumResponse"
              },
              {
                "type" : "null"
              }
            ]
          }
        }
      },
      "QuorumResponse" : {
        "type" : "object",
        "properties" : {
          "mode" : {
            "type" : "string"
          },
          "numerator" : {
            "type" : [
              "integer",
              "null"
            ],
            "format" : "int32"
          },
          "denominator" : {
            "type" : [
              "integer",
              "null"
            ],
            "format" : "int32"
          },
          "absolute" : {
            "type" : [
              "integer",
              "null"
            ],
            "format" : "int32"
          },
          "reachedAt" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "date-time"
          }
        }
      },
      "UpdateQuorumRequest" : {
        "type" : "object",
        "properties" : {
          "mode" : {
            "type" : "string",
            "enum" : [
              "NONE",
              "FRACTION",
              "ABSOLUTE"
            ]
          },
          "numerator" : {
            "type" : [
              "integer",
              "null"
            ],
            "format" : "int32"
          },
          "denominator" : {
            "type" : [
              "integer",
              "null"
            ],
            "format" : "int32"
          },
          "absolute" : {
            "type" : [
              "integer",
              "null"
            ],
            "format" : "int32"
          }
        },
        "required" : [
          "mode"
        ]
      },
      "UpsertTicketTypeRequest" : {
        "type" : "object",
        "properties" : {
          "label" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0
          },
          "maxCapacity" : {
            "type" : [
              "integer",
              "null"
            ],
            "format" : "int32"
          },
          "colorHex" : {
            "type" : "string",
            "pattern" : "^#[0-9a-fA-F]{6}$"
          },
          "position" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [
          "label"
        ]
      },
      "TicketTypeResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "label" : {
            "type" : "string"
          },
          "colorHex" : {
            "type" : "string"
          },
          "maxCapacity" : {
            "type" : [
              "integer",
              "null"
            ],
            "format" : "int32"
          },
          "confirmedCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "position" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "EmailFeedbackEventDto" : {
        "type" : "object",
        "properties" : {
          "recipient" : {
            "type" : "string",
            "minLength" : 1
          },
          "type" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [
          "recipient",
          "type"
        ]
      },
      "EmailFeedbackWebhookRequest" : {
        "type" : "object",
        "properties" : {
          "events" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EmailFeedbackEventDto"
            }
          }
        }
      },
...
      "CreateEventRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "minLength" : 1
          },
          "description" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "startDateTime" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "date-time"
          },
          "endDateTime" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "date-time"
          },
          "timezone" : {
            "type" : "string",
            "minLength" : 1
          },
          "location" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "maxCapacity" : {
            "type" : [
              "integer",
              "null"
            ],
            "format" : "int32"
          },
          "settings" : {
            "$ref" : "#/components/schemas/EventSettingsDto"
          },
          "invitationChannels" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [
                "EMAIL",
                "WHATSAPP"
              ]
            },
            "uniqueItems" : true
          }
        },
        "required" : [
          "name",
          "timezone"
        ]
      },
...
      "ManualPaymentRequest" : {
        "type" : "object",
        "properties" : {
          "tier" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [
          "tier"
        ]
      },
      "ManualPaymentInstructions" : {
        "type" : "object",
        "properties" : {
          "paymentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "reference" : {
            "type" : "string"
          },
          "tier" : {
            "type" : "string"
          },
          "amountMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "currency" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "payee" : {
            "$ref" : "#/components/schemas/PayeeDetails"
          }
        }
      },
      "PayeeDetails" : {
        "type" : "object",
        "properties" : {
          "payeeName" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "mobileMoneyNumber" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "mobileMoneyOperator" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "bankDetails" : {
            "type" : [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}